# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "PveSphere is a comprehensive web-based management platform for Proxmox VE (PVE) clusters.",
        "title": "PveSphere API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "PveSphere Support",
            "url": "https://github.com/pvesphere/pvesphere",
            "email": "support@pvesphere.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "paths": {
...
        "/api/v1/tasks/cluster": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PVE任务模块"
                ],
                "summary": "获取集群任务列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "集群ID",
                        "name": "cluster_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pvesphere_api_v1.ListClusterTasksResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/tasks/log": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PVE任务模块"
                ],
                "summary": "获取任务日志",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "集群ID",
                        "name": "cluster_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "节点名称",
                        "name": "node_name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "任务UPID",
                        "name": "upid",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "起始行号",
                        "name": "start",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "返回行数",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pvesphere_api_v1.GetTaskLogResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/tasks/node": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PVE任务模块"
                ],
                "summary": "获取节点任务列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "集群ID",
                        "name": "cluster_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "节点名称",
                        "name": "node_name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pvesphere_api_v1.ListNodeTasksResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/tasks/status": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PVE任务模块"
                ],
                "summary": "获取任务状态",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "集群ID",
                        "name": "cluster_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "节点名称",
                        "name": "node_name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "任务UPID",
                        "name": "upid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pvesphere_api_v1.GetTaskStatusResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/tasks/stop": {
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PVE任务模块"
                ],
                "summary": "终止任务",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "集群ID",
                        "name": "cluster_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "节点名称",
                        "name": "node_name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "任务UPID",
                        "name": "upid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pvesphere_api_v1.StopTaskResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "pvesphere_api_v1.ClusterTaskItem": {
            "type": "object",
            "properties": {
                "endtime": {
                    "type": "integer"
                },
                "extra": {
                    "type": "object",
                    "additionalProperties": true
                },
                "id": {
                    "type": "string"
                },
                "node": {
                    "type": "string"
                },
                "starttime": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "upid": {
                    "type": "string"
                },
                "user": {
                    "type": "string"
                }
            }
        },
...
        "pvesphere_api_v1.GetTaskLogResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pvesphere_api_v1.TaskLogItem"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "pvesphere_api_v1.GetTaskStatusResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/pvesphere_api_v1.TaskStatusItem"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "pvesphere_api_v1.ListClusterTasksResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pvesphere_api_v1.ClusterTaskItem"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "pvesphere_api_v1.ListNodeTasksResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pvesphere_api_v1.NodeTaskItem"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "pvesphere_api_v1.NodeTaskItem": {
            "type": "object",
            "properties": {
                "endtime": {
                    "type": "integer"
                },
                "extra": {
                    "type": "object",
                    "additionalProperties": true
                },
                "id": {
                    "type": "string"
                },
                "starttime": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "upid": {
                    "type": "string"
                },
                "user": {
                    "type": "string"
                }
            }
        },
...
        "pvesphere_api_v1.StopTaskResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "pvesphere_api_v1.TaskLogItem": {
            "type": "object",
            "properties": {
                "n": {
                    "description": "行号",
                    "type": "integer"
                },
                "t": {
                    "description": "日志内容",
                    "type": "string"
                }
            }
        },
        "pvesphere_api_v1.TaskStatusItem": {
            "type": "object",
            "properties": {
                "endtime": {
                    "type": "integer"
                },
                "exitstatus": {
                    "description": "可能是string或int"
                },
                "extra": {
                    "type": "object",
                    "additionalProperties": true
                },
                "id": {
                    "type": "string"
                },
                "pid": {
                    "type": "integer"
                },
                "pstart": {
                    "type": "integer"
                },
                "starttime": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "upid": {
                    "type": "string"
                },
                "user": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}